up - bug at bearer token , bug at rating , admin routes and other routes testing in postman done

+++ backend/controller/NormalUser.js
@@ -114,7 +114,7 @@ export const submitRating = async (req, res) => {
                 data: {
                     userId,
                     storeId: parseInt(storeId),
-                    rating: parseInt(rating)
+                    rating:  parseFloat(rating)
                 }
             });
         }
